app/requirements.txt
aws-cdk-lib==2.126.0
constructs>=10.0.0,<11.0.0
boto3>=1.34.39,<2.0.0